Restructured code to use a ffmeta module. Code tested on Python 3 with Flask 1.0. Using SqlAlchemy instead of Flask-SqlAlchemy, to enable testing of the part of the codebase that interacts with the database independently of a Flask Application.
