--- api.config.cfg
@@ -1,15 +1,0 @@
-DEBUG=False
-
-# Metadata flatfile location
-METADATA_FILE="data/FFMetadata20180221.csv"
-
-# Database credentials. The following are just placeholders - replace with real credentials.
-AWS_RDS_USER="db_user"
-AWS_RDS_PASS="db_password"
-AWS_RDS_ENDPOINT="db_hostname"
-AWS_RDS_PORT=3306
-DB_NAME="db_name"
-
-# SQLAlchemy settings
-SQLALCHEMY_DATABASE_URI = "mysql+pymysql://{}:{}@{}:{}/{}".format(AWS_RDS_USER, AWS_RDS_PASS, AWS_RDS_ENDPOINT, AWS_RDS_PORT, DB_NAME)
-SQLALCHEMY_TRACK_MODIFICATIONS = False
