+++ ffmeta/templates/feedback.html
@@ -1,0 +1,13 @@
+{% extends "ffmetadata.html" %}
+{% block title %}Home{% endblock %}
+{% block body %}
+<div class="container">
+    <div class="row">
+        <div class="col-md-2"></div>
+        <div class="col-md-8">
+            <iframe src="https://docs.google.com/forms/d/e/1FAIpQLSc-ip7IQafiGyp3A1LWq9Sx1_s8nbkvUVmdrrC2nqI9Y11jwQ/viewform?embedded=true" width="760" height="1000" frameborder="0" marginheight="0" marginwidth="0" align="center">Loading...</iframe>
+        </div>
+        <div class="col-md-2"></div>
+    </div>
+</div>
+{% endblock %}
