+++ ffmeta/templates/index.html
@@ -1,0 +1,46 @@
+{% extends "ffmetadata.html" %}
+{% block title %}Home{% endblock %}
+{% block body %}
+<div class="container">
+    <div class="row">
+        <div class="col-md-12">
+            <div class="page-header"><h2>Fragile Families and Child Wellbeing Study</h2></div>
+                <div class="container">
+                    <div class="row">
+                        <div class="col-md-6">
+                            <p class="lead">Metadata API</p>
+                            <p>We provide three API endpoints:
+                                <ul>
+                                    <li><p>
+                                        <code>select(varName, fieldName)</code><br/>
+                                        &emsp;Returns metadata for variable <code>varName</code>.<br/>
+                                        &emsp;(Optionally, returns only the field specified by <code>fieldName</code>.)<br/>
+                                        &emsp;e.g. <code><a href="{{ url_for('api.selectMetadata') + '?varName=cm1relf' }}">GET /select?varName=cm1relf</a></code>
+
+                                    </p></li>
+                                    <li><p>
+                                        <code>filter(*fieldNames)</code><br/>
+                                        &emsp;Return a list of variables where <code>fieldName</code> matches the provided value.<br/>
+                                        &emsp;e.g. <code><a href="{{ url_for('api.filterMetadata') + '?responses=sometimes' }}">GET /filter?responses=sometimes</a></code>
+                                    </p></li>
+                                    <li><p>
+                                        <code>search(query, fieldName)</code><br/>
+                                        &emsp;Return a list of variables where <code>query</code> is found in <code>fieldName</code>.<br/>
+                                        &emsp;e.g. <code><a href="{{ url_for('api.searchMetadata') + '?query=age&fieldName=label' }}">GET /search?query=age&fieldName=label</a></code>
+                                    </p></li>
+                                </ul>
+                        </div>
+                        <div class="col-md-6">
+                            <p class="lead">Other utilities for FFCWS data users</p>
+                            <p>To browse variables interactively, click <em>Browse online</em>.</p>
+                            <p>To download metadata for all variables as a CSV file, click <em>Download metadata</em>.</p>
+                            <p>If you have any comments or ideas for how to improve the metadata or this API, click <em>Feedback</em>.</p>
+                            <p>For more information on the Fragile Families and Child Wellbeing Study, click <em>About FFCWS</em>.</p>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+{% endblock %}
