+++ requirements.txt
@@ -1,6 +1,6 @@
 click==6.7
-Flask==0.12.2
-Flask-SQLAlchemy==2.3.2
+Flask==1.0.0
+SQLAlchemy==1.2.4
 itsdangerous==0.24
 Jinja2==2.10
 MarkupSafe==1.0
